extracted login styles into it's own styles file
