+++ src/components/Login/Login.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { View, Text, StyleSheet, Image, StatusBar, TouchableOpacity } from 'react-native';
 import LoginForm from './login_form'
+import styles from '../../../src/styles/login/loginStyles'
 
 export default class Login extends Component {
   static navigationOptions = {
@@ -27,33 +28,3 @@ export default class Login extends Component {
     )
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#16a085'
-  },
-  logoContainer: {
-    alignItems: 'center',
-    flexGrow: 1,
-    justifyContent: 'center'
-  },
-  logo: {
-    width: 100,
-    height: 100
-  },
-  title: {
-    color: '#FFF',
-    marginTop: 10,
-    textAlign: 'center',
-    width: 150,
-    opacity: 0.7,
-    fontSize: 20
-  },
-  header: {
-    color: '#FFF',
-    marginTop: 5,
-    textAlign: 'center',
-    opacity: 0.7
-  }
-})
